improve performance of async with batching
when populating the sorting list make parallel calls to logSources that
have been depleted based on a batch size. The higher the batch the more
lines will be printed out at a time and the higher the chance of a
potential set of parallel calls.

Of course, the worst case scenario is each logSource gets drained one at
a time. This is due to happen stance of ordering, but this is basically
rare assuming on average a log source is an equal candidate for the next
sequential message.